+++ src/PhpSpreadsheet/Calculation/DateTimeExcel/Helpers.php
@@ -19,6 +19,8 @@ class Helpers
      */
     public static function isLeapYear(int|string $year): bool
     {
+        $year = (int) $year;
+
         return (($year % 4) === 0) && (($year % 100) !== 0) || (($year % 400) === 0);
     }