+++ src/PhpSpreadsheet/Helper/Html.php
@@ -693,7 +693,7 @@ class Html
     {
         preg_match_all('/\d+/', $rgbValue, $values);
         foreach ($values[0] as &$value) {
-            $value = str_pad(dechex($value), 2, '0', STR_PAD_LEFT);
+            $value = str_pad(dechex($value), 2, '0', STR_PAD_LEFT); // @phpstan-ignore-line
         }
 
         return implode('', $values[0]);