+++ src/PhpSpreadsheet/Writer/Xls/Worksheet.php
@@ -2374,7 +2374,7 @@ class Worksheet extends BIFFwriter
         }
 
         // Slurp the file into a string.
-        $data = (string) fread($bmp_fd, (int) filesize($bitmap));
+        $data = (string) fread($bmp_fd, (int) filesize($bitmap)); // @phpstan-ignore-line
 
         // Check that the file is big enough to be a bitmap.
         if (strlen($data) <= 0x36) {